Fix race condition in react/whenever based signal stress test
On MoarVM signal() uses await to wait for the signal handler to be installed
(which happens on the event loop thread). The react/whenever setup code
however installs its own $*AWAITER, so an await in one whenever's tap handler
does not block the rest of the setup code. Instead, those handlers will be
processed further at the end of the setup. The racy test on the other hand
relied on the whenever handlers to be installed in sequence as part of the
setup block messaged the parent process readiness for receiving signals.

Fix by sending that message from a whenever handler itself so we can be
absolutely sure that the setup is finished.